+++ src/main/java/test/shop/application/service/item/ItemService.java
@@ -59,7 +59,7 @@ public class ItemService {
        item.reorderImage(imageId, newPosition);
     }
 
-    public Page<ProductDto> findItems(int page, int size, Range<Integer> range, String category, int ratings) {
+    public Page<ProductDto> findItems(int page, int size, Range<Integer> range, String category, double ratings) {
         Map<String, Object> params = new HashMap<>();
         params.put("category", category);
         params.put("ratings", ratings);
